+++ esp-storage/build.rs
@@ -1,4 +1,13 @@
+use esp_metadata_generated::Chip;
+
 fn main() -> Result<(), String> {
+    if !cfg!(feature = "emulation") {
+        // Load the configuration file for the configured device:
+        let chip = Chip::from_cargo_feature()?;
+
+        // Define all necessary configuration symbols for the configured device:
+        chip.define_cfgs();
+    }
     if cfg!(feature = "esp32") {
         match std::env::var("OPT_LEVEL") {
             Ok(level) if std::env::var("CI").is_err() => {