+++ esp-storage/src/common.rs
@@ -1,19 +1,36 @@
 use core::mem::MaybeUninit;
 
 use crate::chip_specific;
+#[cfg(multi_core)]
+use crate::multi_core::MultiCoreStrategy;
 
-#[derive(Debug)]
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
 #[non_exhaustive]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+/// Flash storage error.
 pub enum FlashStorageError {
+    /// I/O error.
     IoError,
+    /// I/O operation timed out.
     IoTimeout,
+    /// Flash could not be unlocked for writing.
     CantUnlock,
+    /// Address or length not aligned to required boundary.
     NotAligned,
+    /// Address or length out of bounds.
     OutOfBounds,
+    /// Cannot write to flash as more than one core is running.
+    /// Either manually suspend the other core, or use one of the available strategies:
+    /// * [`FlashStorage::multicore_auto_park`]
+    /// * [`FlashStorage::multicore_ignore`]
+    #[cfg(multi_core)]
+    OtherCoreRunning,
+    /// Other error with the given error code.
     Other(i32),
 }
 
 #[inline(always)]
+/// Check return code from flash operations.
 pub fn check_rc(rc: i32) -> Result<(), FlashStorageError> {
     match rc {
         0 => Ok(()),
@@ -24,9 +41,13 @@ pub fn check_rc(rc: i32) -> Result<(), FlashStorageError> {
 }
 
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+/// Flash storage abstraction.
 pub struct FlashStorage {
     pub(crate) capacity: usize,
     unlocked: bool,
+    #[cfg(multi_core)]
+    pub(crate) multi_core_strategy: MultiCoreStrategy,
 }
 
 impl Default for FlashStorage {
@@ -36,13 +57,18 @@ impl Default for FlashStorage {
 }
 
 impl FlashStorage {
+    /// Flash word size in bytes.
     pub const WORD_SIZE: u32 = 4;
+    /// Flash sector size in bytes.
     pub const SECTOR_SIZE: u32 = 4096;
 
+    /// Create a new flash storage instance.
     pub fn new() -> FlashStorage {
         let mut storage = FlashStorage {
             capacity: 0,
             unlocked: false,
+            #[cfg(multi_core)]
+            multi_core_strategy: MultiCoreStrategy::Error,
         };
 
         #[cfg(not(any(feature = "esp32", feature = "esp32s2")))]
@@ -112,9 +138,16 @@ impl FlashStorage {
     }
 
     pub(crate) fn internal_erase(&mut self, sector: u32) -> Result<(), FlashStorageError> {
-        self.unlock_once()?;
+        #[cfg(multi_core)]
+        let unpark = self.multi_core_strategy.pre_write()?;
 
-        check_rc(chip_specific::spiflash_erase_sector(sector))
+        self.unlock_once()?;
+        check_rc(chip_specific::spiflash_erase_sector(sector))?;
+
+        #[cfg(multi_core)]
+        self.multi_core_strategy.post_write(unpark);
+
+        Ok(())
     }
 
     pub(crate) fn internal_write(
@@ -122,12 +155,19 @@ impl FlashStorage {
         offset: u32,
         bytes: &[u8],
     ) -> Result<(), FlashStorageError> {
-        self.unlock_once()?;
+        #[cfg(multi_core)]
+        let unpark = self.multi_core_strategy.pre_write()?;
 
+        self.unlock_once()?;
         check_rc(chip_specific::spiflash_write(
             offset,
             bytes.as_ptr() as *const u32,
             bytes.len() as u32,
-        ))
+        ))?;
+
+        #[cfg(multi_core)]
+        self.multi_core_strategy.post_write(unpark);
+
+        Ok(())
     }
 }