+++ xtask/src/lib.rs
@@ -3,7 +3,7 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use anyhow::{Result, anyhow};
+use anyhow::{Context, Result, anyhow};
 use cargo::CargoAction;
 use esp_metadata::{Chip, Config, TokenStream};
 use serde::{Deserialize, Serialize};
@@ -137,6 +137,7 @@ impl Package {
         false
     }
 
+    /// Does the package have any host tests?
     pub fn has_host_tests(&self, workspace: &Path) -> bool {
         let package_path = workspace.join(self.to_string()).join("src");
 
@@ -291,6 +292,7 @@ impl Package {
             _ => {}
         }
 
+        log::debug!("Features for package '{}': {:?}", self, features);
         features
     }
 
@@ -334,9 +336,13 @@ impl Package {
             Package::EspPreempt => {
                 cases.push(vec!["esp-alloc".to_owned(), "esp-hal/unstable".to_owned()])
             }
+            Package::EspStorage => {
+                cases.push(vec!["defmt".to_owned()]);
+            }
             _ => {}
         }
 
+        log::debug!("Lint feature cases for package '{}': {:?}", self, cases);
         cases
     }
 
@@ -373,6 +379,11 @@ impl Package {
 
     /// Creates a tag string for this [`Package`] combined with a semantic version.
     pub fn tag(&self, version: &semver::Version) -> String {
+        log::debug!(
+            "Creating tag for package '{}' with version '{}'",
+            self,
+            version
+        );
         format!("{self}-v{version}")
     }
 
@@ -527,12 +538,17 @@ pub fn execute_app(
 // ----------------------------------------------------------------------------
 // Helper Functions
 
-// Copy an entire directory recursively.
+/// Copy an entire directory recursively.
 // https://stackoverflow.com/a/65192210
 pub fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> Result<()> {
-    fs::create_dir_all(&dst)?;
+    log::debug!(
+        "Copying directory '{}' to '{}'",
+        src.as_ref().display(),
+        dst.as_ref().display()
+    );
+    fs::create_dir_all(&dst).with_context(|| "Failed to create a {dst}")?;
 
-    for entry in fs::read_dir(src)? {
+    for entry in fs::read_dir(src).with_context(|| "Failed to read {src}")? {
         let entry = entry?;
         let ty = entry.file_type()?;
 
@@ -550,7 +566,7 @@ pub fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> Result<()> 
 /// workspace.
 pub fn package_paths(workspace: &Path) -> Result<Vec<PathBuf>> {
     let mut paths = Vec::new();
-    for entry in fs::read_dir(workspace)? {
+    for entry in fs::read_dir(workspace).context("Failed to read {workspace}")? {
         let entry = entry?;
         if entry.file_type()?.is_dir() && entry.path().join("Cargo.toml").exists() {
             paths.push(entry.path());
@@ -558,6 +574,12 @@ pub fn package_paths(workspace: &Path) -> Result<Vec<PathBuf>> {
     }
 
     paths.sort();
+
+    log::debug!(
+        "Found {} packages in workspace '{}':",
+        paths.len(),
+        workspace.display()
+    );
 
     Ok(paths)
 }
@@ -590,6 +612,7 @@ pub fn format_package(workspace: &Path, package: Package, check: bool) -> Result
     Ok(())
 }
 
+/// Run the host tests for the specified package.
 pub fn run_host_tests(workspace: &Path, package: Package) -> Result<()> {
     log::info!("Running host tests for package: {}", package);
     let package_path = workspace.join(package.as_ref());
@@ -702,11 +725,14 @@ fn format_package_path(workspace: &Path, package_path: &Path, check: bool) -> Re
     ));
     cargo_args.extend(source_files);
 
+    log::debug!("{cargo_args:#?}");
+
     cargo::run(&cargo_args, &package_path)
 }
 
 /// Update the metadata and chip support table in the esp-hal README.
 pub fn update_metadata(workspace: &Path, check: bool) -> Result<()> {
+    log::info!("Updating esp-metadata and chip support table...");
     update_chip_support_table(workspace)?;
     generate_metadata(workspace, save)?;
 
@@ -715,7 +741,8 @@ pub fn update_metadata(workspace: &Path, check: bool) -> Result<()> {
     if check {
         let res = std::process::Command::new("git")
             .args(["diff", "HEAD", "esp-metadata-generated"])
-            .output()?;
+            .output()
+            .context("Failed to run `git diff HEAD esp-metadata-generated`")?;
         if !res.stdout.is_empty() {
             return Err(anyhow::Error::msg(
                 "detected `esp-metadata-generated` changes. Run `cargo xtask update-metadata`, and commit the changes.",
@@ -767,8 +794,10 @@ fn save(out_path: &Path, tokens: TokenStream) -> Result<()> {
 }
 
 fn update_chip_support_table(workspace: &Path) -> Result<()> {
+    log::debug!("Updating chip support table in README.md...");
     let mut output = String::new();
-    let readme = std::fs::read_to_string(workspace.join("esp-hal").join("README.md"))?;
+    let readme = std::fs::read_to_string(workspace.join("esp-hal").join("README.md"))
+        .context("Failed to read {workspace}")?;
 
     let mut in_support_table = false;
     let mut generate_support_table = true;
@@ -803,7 +832,10 @@ fn update_chip_support_table(workspace: &Path) -> Result<()> {
 pub fn find_packages(path: &Path) -> Result<Vec<PathBuf>> {
     let mut packages = Vec::new();
 
-    for result in fs::read_dir(path)? {
+    for result in
+        fs::read_dir(path).with_context(|| format!("Failed to read {}", path.display()))?
+    {
+        log::debug!("Inspecting path: {}", path.display());
         let entry = result?;
         if entry.path().is_file() {
             continue;
@@ -816,6 +848,12 @@ pub fn find_packages(path: &Path) -> Result<Vec<PathBuf>> {
             packages.extend(find_packages(&entry.path())?);
         }
     }
+
+    log::debug!(
+        "Found {} packages in path '{}':",
+        packages.len(),
+        path.display()
+    );
 
     Ok(packages)
 }