+++ esp-storage/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `defmt` feature and `FlashStorage`, `FlashStorageError` now implements `Defmt` (#4127) 
 
 ### Changed
 
@@ -18,6 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [v0.8.0] - 2025-09-10
+
+### Added
+
+- Added strategies for dealing with multi-core systems (#4082)
 
 ## [v0.7.0] - 2025-07-16